RSNNS 0.3-1 fixes for solaris
git-svn-id: svn+ssh://atolon.ugr.es/home/ts/code/trunk/Rpackages/RSNNS@195 4da2758a-f866-4958-9045-a385b76b148e

+++ src/cc_prune.cpp
@@ -92,7 +92,7 @@ float SnnsCLib::cc_calculatePruneError(int prune_func,int p,int n,float sse)
 {
     switch (prune_func){
       case SBC:
-	return n * log(sse/n) + p*log(n);
+	return n * log(sse/n) + p*log((double)n);
       case AIC:
 	return n* log(sse/n) + p*2;
       case CMSEP:
